Run mypy from tox
This avoids the need to duplicate our dependency list in multiple
places and allows us to take advantage of tox's dependency management
infrastructure, to ensure we always get the latest and greatest version
of a package allowed by upper-constraints.

While here, we also update the versions of the remaining pre-commit
hooks and change the indentation of the tox file to align with the two
indent spacing used for other SDK projects. This makes copy-pasting
easier.

Change-Id: Ibde8ecda673b2346c82aab68d4f4b49be08414ae
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>